feat : skill equip ui

+++ Assets/_Code/KMJ/UnitSystem/Unit/SkillSystem/SkillSO.cs
@@ -8,7 +8,7 @@ namespace Code.UnitSystem.SkillSystem
     {
         public UnitType unitType = UnitType.None;
         public string skillName;
-        public float SkillCost;
+        public int SkillCost;
         public Image skillUIImage;
         public string className;
     }
